Merge after compile est

+++ src/main/java/dog/boopr/boopr/controllers/BreedController.java
@@ -14,7 +14,9 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 
 import dog.boopr.boopr.models.Breed;
+import dog.boopr.boopr.models.User;
 import dog.boopr.boopr.repositories.BreedRespository;
+import dog.boopr.boopr.services.UserServices;
 
 
 @RestController
@@ -23,7 +25,13 @@ public class BreedController {
     @Autowired
     private BreedRespository breedDao;
 
-    //Get breeds
+    @Autowired
+    private UserServices userService;
+
+    /**
+     * @return This route returns a list of all breeds with IDs in a JSONArray
+     * @throws JSONException
+     */
     @RequestMapping(value="/api/breeds", method=RequestMethod.GET, produces=MediaType.APPLICATION_JSON_VALUE)
     public String getBreeds() throws JSONException{
 
@@ -46,16 +54,31 @@ public class BreedController {
         return breeds.toString();
     }
 
-    //Delete Breeds
+    /**
+     * Admin permissions only
+     * @param id The Id of the breed you are wanting to delete
+     * @return JSONObject with a message or error depending on success
+     */
     @RequestMapping(value="/api/breed/delete/{id}", method=RequestMethod.POST, produces=MediaType.APPLICATION_JSON_VALUE)
     public String deleteBreed(
         @PathVariable Long id){
             try{
-                Breed breed = breedDao.getOne(id);
-                breedDao.delete(breed);
+
+                User user = userService.getCurrentUser();
+
+                if(userService.userIsRole(user, "admin")){
+
+                    Breed breed = breedDao.getOne(id);
+                    breedDao.delete(breed);
+
+                    JSONObject response = new JSONObject();
+                    response.put("message","Deleted Breed!");
+                    return response.toString();
+                }
                 JSONObject response = new JSONObject();
-                response.put("message","Deleted Breed!");
+                response.put("error","Youre not an admin!");
                 return response.toString();
+                
             }catch(Exception e){
                 e.printStackTrace();
                 return " { 'error' : '" + e.toString() + " ' }";
